Move to_canonic_path() from utils to path.py
