+++ nextline/spawned/utils.py
@@ -1,4 +1,3 @@
-import os
 from types import FrameType
 from typing import Any, Callable, ContextManager, Optional
 
@@ -31,25 +30,3 @@ def WithContext(
         return _create_local_trace()(frame, event, arg)
 
     return _global_trace
-
-
-def ToCanonicPath() -> Callable[[str], str]:
-    # Based on Bdb.canonic()
-    # https://github.com/python/cpython/blob/v3.10.5/Lib/bdb.py#L39-L54
-
-    cache = dict[str, str]()
-
-    def _to_canonic_path(filename: str) -> str:
-        if filename == "<" + filename[1:-1] + ">":
-            return filename
-        canonic = cache.get(filename)
-        if not canonic:
-            canonic = os.path.abspath(filename)
-            canonic = os.path.normcase(canonic)
-            cache[filename] = canonic
-        return canonic
-
-    return _to_canonic_path
-
-
-to_canonic_path = ToCanonicPath()
